backend/requirements.txt
# Core
Django>=4.2,<5
djangorestframework>=3.14,<4
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.3,<5
psycopg2-binary>=2.9
python-dotenv>=1.0

# ML
pandas>=2.1
numpy>=1.24

# Utilities
Pillow>=10.0
requests>=2.31
psutil>=5.9
openai>=1.30
catboost
